implement local and jwt passport strategy with auth guard decorator

+++ src/users/users.controller.ts
@@ -5,7 +5,6 @@ import {
   Delete,
   Get,
   HttpCode,
-  Logger,
   NotFoundException,
   Param,
   ParseUUIDPipe,
@@ -13,12 +12,17 @@ import {
   Post,
   Query,
   SerializeOptions,
+  UnauthorizedException,
+  UseGuards,
   UseInterceptors,
 } from '@nestjs/common';
 import { CreateUserDto } from './input/create-user.dto';
 import { UpdateUserDto } from './input/update-user.dto';
 import { ListUsersFilter } from './input/list-users.filter';
 import { UsersService } from './users.service';
+import { AuthGuardJwt } from 'src/auth/auth-guard.jwt';
+import { CurrentUser } from './current-user.decorator';
+import { IUser } from './user.entity';
 
 @Controller('users')
 @SerializeOptions({ excludePrefixes: ['_'] })
@@ -31,16 +35,33 @@ export class UsersController {
     return this.usersService.findAll(filter);
   }
 
-  @Get(':id')
-  @UseInterceptors(ClassSerializerInterceptor)
-  async find(@Param('id', ParseUUIDPipe) id: string) {
-    const user = this.usersService.findById(id);
+  // @Get(':id')
+  // @UseGuards(AuthGuardJwt)
+  // @UseInterceptors(ClassSerializerInterceptor)
+  // async find(
+  //   @CurrentUser() user: IUser,
+  //   @Param('id', ParseUUIDPipe) id: string,
+  // ) {
+  //   if (user.id !== id) {
+  //     throw new UnauthorizedException();
+  //   }
 
+  //   if (!user) {
+  //     throw new NotFoundException();
+  //   }
+
+  //   return user;
+  // }
+
+  @Get('profile')
+  @UseGuards(AuthGuardJwt)
+  @UseInterceptors(ClassSerializerInterceptor)
+  async getProfile(@CurrentUser() user: IUser) {
     if (!user) {
-      throw new NotFoundException();
+      throw new UnauthorizedException();
     }
 
-    return this.usersService.findById(id);
+    return user;
   }
 
   @Post()
@@ -50,9 +71,16 @@ export class UsersController {
   }
 
   @Patch(':id')
+  @UseGuards(AuthGuardJwt)
   @UseInterceptors(ClassSerializerInterceptor)
-  async update(@Param('id', ParseUUIDPipe) id, @Body() input: UpdateUserDto) {
-    const user = this.usersService.findById(id);
+  async update(
+    @CurrentUser() user: IUser,
+    @Param('id', ParseUUIDPipe) id,
+    @Body() input: UpdateUserDto,
+  ) {
+    if (user.id !== id) {
+      throw new UnauthorizedException();
+    }
 
     if (!user) {
       throw new NotFoundException();
@@ -63,8 +91,13 @@ export class UsersController {
   }
 
   @Delete(':id')
+  @UseGuards(AuthGuardJwt)
   @HttpCode(204)
-  async remove(@Param('id', ParseUUIDPipe) id) {
+  async remove(@CurrentUser() user: IUser, @Param('id', ParseUUIDPipe) id) {
+    if (user.id !== id) {
+      throw new UnauthorizedException();
+    }
+
     this.usersService.delete(id);
   }
 }
